Moved some pure XAML classes into the XAML project to remove clutter in the main assembly.
