+++ src/MarkPad.XAML/packages.config
@@ -1,4 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="Analects.XAMLConverters" version="1.0" />
+  <package id="MahApps.Metro" version="0.4.0.17" />
 </packages>
