+++ src/MarkPad/Document/DocumentView.xaml
@@ -4,9 +4,12 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
     xmlns:System="clr-namespace:System;assembly=mscorlib"
-    xmlns:shared="http://schemas.markpad.net/winfx/xaml/shared" xmlns:cal="http://www.caliburnproject.org"
-    xmlns:awe="clr-namespace:Awesomium.Windows.Controls;assembly=Awesomium.Windows.Controls" xmlns:avalonedit="http://icsharpcode.net/sharpdevelop/avalonedit"
-    xmlns:Document="clr-namespace:MarkPad.Document" xmlns:Framework="clr-namespace:MarkPad.Framework"
+    xmlns:shared="http://schemas.markpad.net/winfx/xaml/shared" 
+    xmlns:controls="http://schemas.markpad.net/winfx/xaml/controls" 
+    xmlns:cal="http://www.caliburnproject.org"
+    xmlns:awe="clr-namespace:Awesomium.Windows.Controls;assembly=Awesomium.Windows.Controls" 
+    xmlns:avalonedit="http://icsharpcode.net/sharpdevelop/avalonedit"
+    xmlns:Document="clr-namespace:MarkPad.Document"
     x:Class="MarkPad.Document.DocumentView"
     mc:Ignorable="d" 
     d:DesignHeight="300" d:DesignWidth="300"
@@ -513,7 +516,7 @@
             <ColumnDefinition Width="{Binding DistractionFree, Converter={shared:BoolToGridLengthConverter Invert=True}}" />
         </Grid.ColumnDefinitions>
 
-        <Framework:FloatingToolBar x:Name="floatingToolBar" PlacementTarget="{Binding ElementName=Editor}" CommandTarget="{Binding ElementName=root}" VerticalOffset="-60" HorizontalOffset="-30">
+        <controls:FloatingToolBar x:Name="floatingToolBar" PlacementTarget="{Binding ElementName=Editor}" CommandTarget="{Binding ElementName=root}" VerticalOffset="-60" HorizontalOffset="-30">
             <Border Background="#f0f0f0" Padding="3,2,3,2" BorderThickness="1" CornerRadius="1" BorderBrush="#999999">
                 <StackPanel Orientation="Horizontal">
                     <Button Width="22"  Command="Document:FormattingCommands.ToggleBold" ToolTip="Bold">B</Button>
@@ -522,7 +525,7 @@
                     <Button Width="78" Command="Document:FormattingCommands.ToggleCodeBlock" ToolTip="Code block">Code block</Button>
                 </StackPanel>
             </Border>
-        </Framework:FloatingToolBar>
+        </controls:FloatingToolBar>
 
         <avalonedit:TextEditor
             x:Name="Editor"
