+++ src/MarkPad/Shell/ShellView.xaml
@@ -5,10 +5,12 @@
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
+    xmlns:s="http://schemas.markpad.net/winfx/xaml/shared"
     xmlns:fw="clr-namespace:MarkPad.Framework"
     xmlns:cm="clr-namespace:Caliburn.Micro;assembly=Caliburn.Micro"
     xmlns:Behaviours="clr-namespace:MahApps.Metro.Behaviours;assembly=MahApps.Metro"
-    xmlns:Shell="clr-namespace:MarkPad.Shell" mc:Ignorable="d" 
+    xmlns:Shell="clr-namespace:MarkPad.Shell" 
+    mc:Ignorable="d" 
     MouseLeftButtonDown="DragMoveWindow"
     Icon="../markpad.ico"
     KeyDown="WindowKeyDown"
@@ -19,55 +21,55 @@
     </i:Interaction.Behaviors>
 
     <i:Interaction.Triggers>
-        <fw:InputBindingTrigger>
-            <fw:InputBindingTrigger.InputBinding>
+        <s:InputBindingTrigger>
+            <s:InputBindingTrigger.InputBinding>
                 <KeyBinding Modifiers="Ctrl" Key="N" />
-            </fw:InputBindingTrigger.InputBinding>
+            </s:InputBindingTrigger.InputBinding>
             <cm:ActionMessage MethodName="NewDocument" />
-        </fw:InputBindingTrigger>
-        <fw:InputBindingTrigger>
-            <fw:InputBindingTrigger.InputBinding>
+        </s:InputBindingTrigger>
+        <s:InputBindingTrigger>
+            <s:InputBindingTrigger.InputBinding>
                 <KeyBinding Modifiers="Ctrl" Key="S" />
-            </fw:InputBindingTrigger.InputBinding>
+            </s:InputBindingTrigger.InputBinding>
             <cm:ActionMessage MethodName="SaveDocument" />
-        </fw:InputBindingTrigger>
-        <fw:InputBindingTrigger>
-            <fw:InputBindingTrigger.InputBinding>
+        </s:InputBindingTrigger>
+        <s:InputBindingTrigger>
+            <s:InputBindingTrigger.InputBinding>
                 <KeyBinding Modifiers="Ctrl+Shift" Key="S" />
-            </fw:InputBindingTrigger.InputBinding>
+            </s:InputBindingTrigger.InputBinding>
             <cm:ActionMessage MethodName="SaveAllDocuments" />
-        </fw:InputBindingTrigger>
+        </s:InputBindingTrigger>
 
-        <fw:InputBindingTrigger>
-            <fw:InputBindingTrigger.InputBinding>
+        <s:InputBindingTrigger>
+            <s:InputBindingTrigger.InputBinding>
                 <KeyBinding Modifiers="Ctrl" Key="B" />
-            </fw:InputBindingTrigger.InputBinding>
+            </s:InputBindingTrigger.InputBinding>
             <cm:ActionMessage MethodName="ToggleBold" />
-        </fw:InputBindingTrigger>
-        <fw:InputBindingTrigger>
-            <fw:InputBindingTrigger.InputBinding>
+        </s:InputBindingTrigger>
+        <s:InputBindingTrigger>
+            <s:InputBindingTrigger.InputBinding>
                 <KeyBinding Modifiers="Ctrl" Key="I" />
-            </fw:InputBindingTrigger.InputBinding>
+            </s:InputBindingTrigger.InputBinding>
             <cm:ActionMessage MethodName="ToggleItalic" />
-        </fw:InputBindingTrigger>
-        <fw:InputBindingTrigger>
-            <fw:InputBindingTrigger.InputBinding>
+        </s:InputBindingTrigger>
+        <s:InputBindingTrigger>
+            <s:InputBindingTrigger.InputBinding>
                 <KeyBinding Modifiers="Ctrl" Key="J" />
-            </fw:InputBindingTrigger.InputBinding>
+            </s:InputBindingTrigger.InputBinding>
             <cm:ActionMessage MethodName="ToggleCode" />
-        </fw:InputBindingTrigger>
-        <fw:InputBindingTrigger>
-            <fw:InputBindingTrigger.InputBinding>
+        </s:InputBindingTrigger>
+        <s:InputBindingTrigger>
+            <s:InputBindingTrigger.InputBinding>
                 <KeyBinding Key="F10" />
-            </fw:InputBindingTrigger.InputBinding>
+            </s:InputBindingTrigger.InputBinding>
             <cm:ActionMessage MethodName="ToggleWebView" />
-        </fw:InputBindingTrigger>
-        <fw:InputBindingTrigger>
-            <fw:InputBindingTrigger.InputBinding>
+        </s:InputBindingTrigger>
+        <s:InputBindingTrigger>
+            <s:InputBindingTrigger.InputBinding>
                 <KeyBinding Modifiers="Ctrl" Key="O" />
-            </fw:InputBindingTrigger.InputBinding>
+            </s:InputBindingTrigger.InputBinding>
             <cm:ActionMessage MethodName="OpenDocument" />
-        </fw:InputBindingTrigger>
+        </s:InputBindingTrigger>
     </i:Interaction.Triggers>
 
     <Window.Resources>
@@ -124,18 +126,18 @@
                     <TextBlock TextWrapping="Wrap" FontWeight="Bold" FontSize="10.667">FORMAT TEXT</TextBlock>
                     <TextBlock TextWrapping="Wrap" Opacity="0.75" FontWeight="Bold" FontSize="10.667"><Run Text="HEADERS"/></TextBlock>
                     <TextBlock TextWrapping="Wrap" FontFamily="Courier New">
-				<Run># This is an &lt;h1&gt; tag</Run><LineBreak />
-				<Run>## This is an &lt;h2&gt; tag</Run><LineBreak />
-				<Run>####### This is an &lt;h6&gt; tag</Run>
+                <Run># This is an &lt;h1&gt; tag</Run><LineBreak />
+                <Run>## This is an &lt;h2&gt; tag</Run><LineBreak />
+                <Run>####### This is an &lt;h6&gt; tag</Run>
                     </TextBlock>
 
                     <TextBlock TextWrapping="Wrap" Margin="0,10,0,0" Opacity="0.75" FontSize="10.667"><Run FontWeight="Bold" Text="TEXT STYLE"/></TextBlock>
                     <TextBlock TextWrapping="Wrap" FontFamily="Courier New" >
-				<Run>*This text will be italic*</Run><LineBreak />
-				<Run>_This will also be italic_</Run><LineBreak />
-				<Run>**This text will be bold**</Run><LineBreak />
-				<Run>__This will also be bold__</Run><LineBreak /><LineBreak />
-				<Run>*You **can** combine them*</Run>
+                <Run>*This text will be italic*</Run><LineBreak />
+                <Run>_This will also be italic_</Run><LineBreak />
+                <Run>**This text will be bold**</Run><LineBreak />
+                <Run>__This will also be bold__</Run><LineBreak /><LineBreak />
+                <Run>*You **can** combine them*</Run>
                     </TextBlock>
 
                 </StackPanel>
@@ -192,8 +194,8 @@
             </Rectangle>
 
             <ContentControl x:Name="MDI"
-				Margin="10,0"
-				Grid.Row="1" />
+                Margin="10,0"
+                Grid.Row="1" />
         </Grid>
     </Grid>
 </Window>
